fixes and readme updated

+++ test.h
@@ -5,7 +5,7 @@
 
 #define N_TEST 154
 
-extern uint8_t y_test[N_TEST];
+extern uint16_t y_test[N_TEST];
 extern fixed X_test[N_TEST][N_FEATURE];
-
 #endif
+
